% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{A RC class for multiple linear regression}
\description{
A RC class for multiple linear regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{The formula object containing depedent and independent variables}

\item{\code{data}}{A data frame object to apply the multiple linear regression to

a<-linreg$new(formula=Petal.Length ~ Sepal.Width + Sepal.Length,data=iris)
plot(linreg$new(formula=Petal.Length ~ Sepal.Width + Sepal.Length,data=iris))}
}}

