% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg}
\alias{ridgereg}
\title{A RC class for ridgereg regression}
\description{
A RC class for ridgereg regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{The formula object containing depedent and independent variables}

\item{\code{data}}{A data frame object to apply the multiple linear regression to}

\item{\code{lambda}}{numeric value which is default set as Zer0
a<-ridgereg$new(formula=Petal.Length ~ Sepal.Width + Sepal.Length,data=iris)
plot(ridgereg$new(formula=Petal.Length ~ Sepal.Width + Sepal.Length,data=iris))}
}}


