% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize ariport Flight delays}
\usage{
visualize_airport_delays()
}
\description{
In this you can see that which flight is delay at what is its mean by its lat lng
}
\details{
See flight visualization and airport on graph plot
}
\examples{
visualize_airport_delays()
}
