% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression}
\description{
RC class mainpluate linear algeric caluclautions giving formula and data.
}
\section{Fields}{

\describe{
\item{\code{formula}}{Formula i.e Petal.Length~Sepal.Width+Sepal.Length}

\item{\code{data}}{A data frame iris or faithfull}
}}

\examples{
data(iris)
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$print()
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$pred()
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$summary()
}
