% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg}
\alias{ridgereg}
\title{A RC class for ridgereg regression}
\description{
A RC class for ridgereg regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{The formula object containing depedent and independent variables}

\item{\code{data}}{A data frame object to apply the multiple linear regression to}

\item{\code{lambda}}{numeric value which is default set as Zer0
a<-ridgereg$new(formula=Petal.Length ~ Sepal.Width + Sepal.Length,data=iris)
plot(ridgereg$new(formula=Petal.Length ~ Sepal.Width + Sepal.Length,data=iris))}
}}

\section{Methods}{

\describe{
\item{\code{coeff()}}{Returns a named vector of the regression coefficients in the linear model.}

\item{\code{initialize(formula, data, lambda)}}{From inputs formula and data, generates all the other fields upon object initialization.}

\item{\code{predict(newframe = NULL)}}{Returns the predicted values of the linear model. Can also take a new NAMED data.frame
with covariate values and make new prediction based on them. New NAMED data frame must have same
column order as the data frame in field data.}

\item{\code{print()}}{Gives a printout of the call as well as the calculated regression coefficients.}
}}

